Moves paginated datatables to their own app
