+++ app/grandchallenge/algorithms/templates/algorithms/job_list.html
@@ -1,11 +1,6 @@
-{% extends "base.html" %}
+{% extends "datatables/list_base.html" %}
 {% load url %}
-{% load profiles %}
-{% load workstations %}
 {% load guardian_tags %}
-{% load pathlib %}
-{% load humanize %}
-{% load evaluation_extras %}
 
 {% block title %}
     Algorithm Results - {{ block.super }}
@@ -13,15 +8,9 @@
 
 {% block breadcrumbs %}
     <ol class="breadcrumb">
-        <li class="breadcrumb-item"><a
-                href="{% url 'algorithms:list' %}">Algorithms</a>
-        </li>
-        <li class="breadcrumb-item"><a
-                href="{{ algorithm.get_absolute_url }}">{{ algorithm.title }}
-        </a>
-        <li class="breadcrumb-item active"
-            aria-current="page">Results
-        </li>
+        <li class="breadcrumb-item"><a href="{% url 'algorithms:list' %}">Algorithms</a></li>
+        <li class="breadcrumb-item"><a href="{{ algorithm.get_absolute_url }}">{{ algorithm.title }}</a>
+        <li class="breadcrumb-item active" aria-current="page">Results</li>
     </ol>
 {% endblock %}
 
@@ -39,28 +28,7 @@
         </p>
     {% endif %}
 
-    <div class="table-responsive">
-        <table class="table table-hover table-borderless table-sm">
-            <thead class="thead-light">
-            <tr>
-                {% for column in columns %}
-                    <th {% if column.identifier %}id="{{ column.identifier }}"{% endif %}
-                        {% if column.classes %}class="{% for cls in column.classes %}{{ cls }} {% endfor %}"{% endif %}
-                    >{{ column.title }}</th>
-                {% endfor %}
-            </tr>
-            </thead>
-            <tbody>
-            <tr>
-                <td colspan="100%" class="text-center">
-                    <div class="text-secondary text-center spinner-border" role="status">
-                        <span class="sr-only">Loading Results...</span>
-                    </div>
-                </td>
-            </tr>
-            </tbody>
-        </table>
-    </div>
+    {{ block.super }}
 
     <div class="modal fade" id="resultInfoModal" tabindex="-1" role="dialog"
          aria-labelledby="resultInfoModalTitle" aria-hidden="true">
@@ -88,21 +56,6 @@
 
 {% block script %}
     {{ block.super }}
-
-    <script type="text/javascript">
-        $(document).ready(function () {
-            $('.table').DataTable({
-                order: [[0, "desc"]],
-                lengthChange: false,
-                pageLength: 25,
-                serverSide: true,
-                ajax: {
-                    url: "",
-                },
-                ordering: true,
-            });
-        });
-    </script>
 
     <script type="text/javascript">
         $('#resultInfoModal').on('show.bs.modal', function (event) {
