+++ app/grandchallenge/algorithms/templates/algorithms/job_list_row.html
