+++ app/grandchallenge/datatables/templates/datatables/list_base.html
@@ -1,0 +1,25 @@
+{% extends "base.html" %}
+
+{% block content %}
+    {% include "datatables/partials/datatable.html" with columns=columns %}
+
+{% endblock %}
+
+{% block script %}
+    {{ block.super }}
+
+    <script type="text/javascript">
+        $(document).ready(function () {
+            $('.table').DataTable({
+                order: [[0, "desc"]],
+                lengthChange: false,
+                pageLength: 25,
+                serverSide: true,
+                ajax: {
+                    url: "",
+                },
+                ordering: true,
+            });
+        });
+    </script>
+{% endblock %}
