+++ app/grandchallenge/evaluation/templates/evaluation/leaderboard_detail.html
@@ -6,14 +6,10 @@
 
 {% block breadcrumbs %}
     <ol class="breadcrumb">
-        <li class="breadcrumb-item"><a
-                href="{% url 'challenges:list' %}">Challenges</a>
-        </li>
+        <li class="breadcrumb-item"><a href="{% url 'challenges:list' %}">Challenges</a></li>
         <li class="breadcrumb-item"><a
                 href="{{ challenge.get_absolute_url }}">{% firstof challenge.title challenge.short_name %}</a></li>
-        <li class="breadcrumb-item active">
-            {{ phase.title }} Leaderboard
-        </li>
+        <li class="breadcrumb-item active">{{ phase.title }} Leaderboard</li>
     </ol>
 {% endblock %}
 
@@ -27,29 +23,7 @@
         </div>
     {% endif %}
 
-    <div class="table-responsive mb-3">
-        <table class="table table-hover table-borderless table-sm"
-               id="resultsTable">
-            <thead class="thead-light">
-            <tr>
-                {% for column in columns %}
-                    <th {% if column.identifier %}id="{{ column.identifier }}"{% endif %}
-                        {% if column.classes %}class="{% for cls in column.classes %}{{ cls }} {% endfor %}"{% endif %}
-                    >{{ column.title }}</th>
-                {% endfor %}
-            </tr>
-            </thead>
-            <tbody>
-            <tr>
-                <td colspan="100%" class="text-center">
-                    <div class="text-secondary spinner-border" role="status">
-                        <span class="sr-only">Loading Results...</span>
-                    </div>
-                </td>
-            </tr>
-            </tbody>
-        </table>
-    </div>
+    {% include "datatables/partials/datatable.html" with columns=columns %}
 
     {% if phase.result_display_choice == phase.BEST %}
         <p class="small ml-3">
